+++ CodeConverter/VB/CSToVBProjectContentsConverter.cs
@@ -45,7 +45,7 @@ internal class CSToVBProjectContentsConverter : IProjectContentsConverter
     public async Task InitializeSourceAsync(Project project)
     {
         // TODO: Don't throw away solution-wide effects - write them to referencing files, and use in conversion of any other projects being converted at the same time.
-        project = await ClashingMemberRenamer.RenameClashingSymbolsAsync(project);
+        project = await ClashingMemberRenamer.RenameClashingSymbolsAsync(project, _cancellationToken);
         _convertedVbProject = project.ToProjectFromAnyOptions(_vbCompilationOptions, _vbParseOptions);
         _vbReferenceProject = project.CreateReferenceOnlyProjectFromAnyOptions(_vbCompilationOptions, _vbParseOptions);
         _vbViewOfCsSymbols = (VisualBasicCompilation)await _vbReferenceProject.GetCompilationAsync(_cancellationToken);