+++ CodeConverter/CSharp/BuiltInVisualBasicOperatorSubstitutions.cs
@@ -40,7 +40,7 @@ internal static class VbOperatorConversion
             }
                 
             var notted =
-                node.IsKind(VBasic.SyntaxKind.IsNotExpression, VBasic.SyntaxKind.NotEqualsExpression) ||
+                node.IsKind(VBasic.SyntaxKind.IsNotExpression, VBasic.SyntaxKind.NotEqualsExpression) ^
                 negateExpression;
             var isReferenceComparison = node.IsKind(VBasic.SyntaxKind.IsExpression, VBasic.SyntaxKind.IsNotExpression);