+++ CHANGELOG.md
@@ -7,6 +7,18 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ### Vsix
 
+
+### VB -> C#
+
+
+### C# -> VB
+
+
+## [9.2.6] - 2024-07-08
+
+
+### Vsix
+
 * Compatible with Visual Studio 2022 17.1 onwards
 
 ### VB -> C#